+++ tests/testthat/test-wide2long_omv.R
@@ -1,5 +1,26 @@
 test_that("wide2long_omv works", {
-    # rather simple data set ==================================================================================================================================
+    # rather simple data set ("." as varSep) ==================================================================================================================
+    set.seed(1)
+    dtaTmp <- cbind(data.frame(Year = 1900:2020), as.data.frame(matrix(runif(121 * 12, 0, 100), nrow = 121, dimnames = list(1:121, paste0("X.", month.abb[1:12])))))
+    for (i in 1:12) attr(dtaTmp[[i + 1]], "jmv-desc") <- paste0("Test variable (Month: ", month.abb[i], ")")
+    nmeInp <- tempfile(fileext = ".rds")
+    nmeOut <- tempfile(fileext = "_L.omv")
+    saveRDS(dtaTmp, nmeInp)
+
+    expect_null(wide2long_omv(dtaInp = nmeInp, fleOut = nmeOut, varLst = setdiff(names(dtaTmp), "Year"), varID = "Year", varTme = "Month", varSep = "."))
+    df4Chk <- read_omv(nmeOut)
+    expect_s3_class(df4Chk, "data.frame")
+    expect_equal(dim(df4Chk), c(1452, 3))
+    expect_equal(as.vector(sapply(df4Chk, typeof)), c("integer", "integer", "double"))
+    expect_equal(names(df4Chk), c("Year", "Month", "X"))
+    expect_equal(names(attributes(df4Chk)), c("names", "row.names", "class", "removedRows", "addedRows", "transforms"))
+    expect_equal(names(attributes(df4Chk[[3]])), c("jmv-desc", "name", "id", "columnType", "dataType", "measureType", "formula", "formulaMessage",
+                                                   "parentId", "width", "type", "importName", "description", "transform", "edits", "missingValues"))
+    expect_equal(attr(df4Chk[[3]], "jmv-desc"), "Test variable")
+    expect_equal(c(mean(df4Chk[[3]]), sd(df4Chk[[3]])), c(49.33121, 28.93480), tolerance = 1e-4)
+    unlink(nmeOut)
+
+    # rather simple data set ("_" as varSep) ==================================================================================================================
     set.seed(1)
     dtaTmp <- cbind(data.frame(Year = 1900:2020), as.data.frame(matrix(runif(121 * 12, 0, 100), nrow = 121, dimnames = list(1:121, paste0("X_", month.abb[1:12])))))
     for (i in 1:12) attr(dtaTmp[[i + 1]], "jmv-desc") <- paste0("Test variable (Month: ", month.abb[i], ")")
@@ -45,7 +66,6 @@ test_that("wide2long_omv works", {
     expect_equal(dim(df4Chk), c(1452, 3))
     expect_false(is.unsorted(df4Chk[["Year"]]))
     expect_true(all(df4Chk[["Month"]] == rep(sort(month.abb), length(unique(df4Chk[["Year"]])))))
-
 
     # test cases for code coverage ============================================================================================================================
     expect_error(wide2long_omv(fleInp = nmeInp, varID = "Year", varTme = "Month", varSep = "_"), regexp = "Please use the argument dtaInp instead of fleInp\\.")